backend/frame/requirements.txt
asgiref==3.3.4
Django==3.2
django-background-tasks==1.2.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
html2text==2020.1.16
python-dotenv==0.17.0
pytz==2021.1
sqlparse==0.4.1
web3==5.23.1
